kernel: pass loadaddr for building uImage
On newer kernels (4.x ?) kernel build fails when building an uImage
unless you provide LOADADDR as kernel make var, which is put in the
resulting uImage header. You could append it to
PTXCONF_KERNEL_EXTRA_MAKEVARS but this would be just a workaround. This
patch extends the menu if uImage is selected so you can directly input
the desired address.

Signed-off-by: Alexander Dahl <post@lespocky.de>