image_kernel: new options for "appended DT" images
On a platform where we can not load a separate device tree binary blob
with the old bootloader we use the kernel option
CONFIG_ARM_APPENDED_DTB to load the DT blob. This requires creating a
zImage, concatenate the dtb file and put this back to an uImage. This
patch extends the image generation rules so this is possible
automatically with just a few options in the menu.

Signed-off-by: Alexander Dahl <post@lespocky.de>