host-imx-cst: Add PKCS#11 support
This adds PKCS#11 support to the CST. PKCS#11 support is not supported
natively by the CST so we have to patch it. The CST is provided in
binary form, but the tarball contains a header file and a static library
which can be used to extend the OpenSSL capabilities. Previously we
didn't have to recompile the cst binary, so the necessary compilation
step is provided as part of this patch.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>