pkg-config-wrapper: remove sysroot only from specified variables
The PTXDIST_PKG_CONFIG_VAR_NO_SYSROOT variable allows to specify
pkg-config variable that shall be relative to the target rootfs and not
contain the sysroot path. This is important for variables that are used
for writing files to the target rootfs, e.g., systemduserunitdir.

Currently, the sysroot is removed from all pkg-config variables, if
PTXDIST_PKG_CONFIG_VAR_NO_SYSROOT is set, but packages already specify
the name of the variable that must not include the sysroot.

Add a check to remove the sysroot only from variables that are actually
contained in PTXDIST_PKG_CONFIG_VAR_NO_SYSROOT.

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>