libnl: move to staging (superseded by libnl3)
The version 1 of libnl is long outdated and deprecated, with the
last release 1.1.4 from 2013.

Inside ptxdist, iw 3.14 was the last official user of LIBNL, all
other packages use LIBNL3. With the update to iw 5.3, this was
switched to LIBNL3, too.
All other packages already switched to libnl version 3 (ptxdist rule
"libnl3") for a long time, and so is the year-old recommendation of
libnl mainline.

Therefore move libnl to staging, and schedule for removal.

Signed-off-by: Andreas Pretzsch <apr@cn-eng.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>