image-kernel-fit: Add package
This adds a package which generates a FIT image from the kernel image
and all selected device trees in PTXCONF_DTC_OFTREE_DTS. The image will
be signed if desired.

** Note **
Currently the U-Boot mkimage tool supports PKCS#11, but there is no way
to specify the PIN in an environment variable like it is done in many
other tools. This means you will be prompted for a PIN each time a
signed FIT image is being built.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>