alsa-lib: provide a way to limit the PCM modules
On some platforms it makes sense to limit the available PCM modules to
the really required ones (application specific) or useful ones (hardware
specific). This change enables the user to select specific modules which
will be later on present in the run-time system.

To be backward compatible, the default keyword 'all' will still build all
available PCM modules. The list of available PCM modules is part of the
package's 'configure' help text.

Signed-off-by: Juergen Borleis <jbe@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>