ptxd_lib_dgen: handle missing config files better
The prepare or image stages used to have a dependency on the config files
in the past. Changes in the config file are now handled by the pkghash so
the dependency was removed. Howerver, the dependency was often used to
provide an extra features:
Packages provide a target for the config to print a custom error message if
the config file was missing.
Reintroduce the dependency again but only if the the file is missing to
restore this feature.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>