at91bootstrap2: Revise make env and opt
The 'ARCH=$(PTXCONF_BAREBOX_ARCH_STRING)' is present since first version
of this package and maybe a leftover from some other package this is
based on. However that ARCH variable is not used by the at91boostrap 3.x
Makefile and barebox might not even be enabled.

The CROSS_COMPILE variable is changed to what other bootloader packages
use, and we support passing V=1 now.

Signed-off-by: Alexander Dahl <ada@thorsis.com>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>