config: boardsetup: add nfs client specifications
This patch adds additional client specifications for the nfsroot mount.
The default value is the same as it was hardcoded before. This offers to
change this line without maintaining the whole file in a separate BSP.

The client specifications is part of the official export format which is
set right after the directory which is exported, for more information
lookup:

https://linux.die.net/man/8/unfsd

Signed-off-by: Alexander Aring <alex.aring@gmail.com>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>