python3-numpy: avoid host search paths
If no site.cfg is defined, then numpy will search in /usr and /usr/local
for various headers and libraries. As a result, those paths may be added to
the compiler command-line and building fails with 'Bad search path' errors.

Create a site.cfg with empty library_dirs and include_dirs to avoid this.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>