python3: add option to include bundled pip
Python can be configured --with-ensurepip, which bundles an included pip
installer. If the option is enabled, the build process also generates
two wheel files for pip and setuptools:

  Lib/ensurepip/_bundled/pip-19.0.3-py2.py3-none-any.whl
  Lib/ensurepip/_bundled/setuptools-40.8.0-py2.py3-none-any.whl

which need to be installed on targetinstall.

Now that 'python3 -m venv env' works, it turns out that the 'activate'
script was not installed, which is fixed in this patch as well.

Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>