ptxd_get_path: don't use 'eval'
It breaks paths with '\', whitespace, etc. bacause the quoting is gone when
'ls' is called.
This means that no globing is expanded for the arguments. That should not
be a problem, because that should happen when ptxd_get_path is called.

Now IFS must be a newline because arguments can contain white-spaces
now. This ensures that the array is setup correctly.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>