fcgi: version bump 2.4.0 -> 2.4.2
The repository on sourceforge doesn't exist any more, there seems to be
a maintained repo on github, so switch to it.

The patches from 2.4.0 can be removed: 0001, 0002 and 0004 are already
in upstream, 0003 isn't necessary any more as the file doesn't contain
an EOF (which was the reason to include stdio.h).

The new upstream repo doesn't contain configure, so we have to autogen it.

Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>
[add patch description, check patches, remove patches]
Signed-off-by: Patrick Enns <patrick.enns@wago.com>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>