unfs3: fix attribute setting for symlinks
unfs makes use of utime() which is unsuitable to set last access and
modification times for symlinks. So make use of the O_PATH flag to
open() and use the futimens() function.

As a side effect this is less racy and improves resolution from seconds
to nanoseconds.

Note this will break builds of unfs on Windows. The futimens() function
is specified in POSIX.1-2008, so this shouldn't be a problem for other
platforms.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Message-Id: <20200218145720.4992-1-u.kleine-koenig@pengutronix.de>
[mol: fix path author as requested]
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>