unfs3: fix attribute settings, second try
The last change to unfs3 repaired some things for symlinks but in return
broke stuff for regular files.

With this change the previous attempt is replaced by something better
tested now.

Instead of letting unfs open the file to adapt, use fchmodat() et al,
which are racy if the underlaying filesystem changes, but on the nfs
side it is fine.

Fixes: 4e21b490eb28 ("unfs3: fix attribute setting for symlinks")
Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
[mol: simplify patch]
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>