dropbear: version bump 2017.75 -> 2019.78
The major change with 2018.76 was in configuration, moving from editing
'options.h' to adding things to 'localoptions.h'. The latter file must
be created in the extracted sources, a pragmatic approach was used to
add the defines, as seen in the buildroot project.

All options were compared to 'default_options.h' and 'sysoptions.h' from
the dropbear source. Not all options are in default_options, e.g. the
deprecated md5 hmac. A warning for that one was added.

Various things were fixed and improved in dropbear, changes notable:

+ 2019.78 / 2019.77
  - Fix 2018.76 regressions
  - Improve failure delay consistency
  - Change handling of failed authentication to avoid disclosing valid
    usernames, CVE-2018-15599.

+ 2018.76
  - Enable hardening build flags
  - New key exchange options
  - Improved DSS and RSA pubkey validation
  - Fix null pointer crash on malformed keys
  - Update curve25519-donna implementation

Signed-off-by: Alexander Dahl <ada@thorsis.com>
Message-Id: <20200311132408.11924-3-ada@thorsis.com>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>