qt5: drop qtdeclarative opengl dependency
Since 5.14 the qtdeclarative can be build with -no-opengl option so we
can remove the dependency. Now all 3D modules need to add the opengl
dependency by there own. We also must ensure that all 3D modules are
skipped if -no-opengl is specified.

The particles module depends on the opengl support and the module isn't
build if the support is missing. So we need a way to handle this too.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>
Message-Id: <20200312113516.21154-2-m.felsch@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>