nodejs: version bump 6.9.5 -> 12.14.1 LTS
We also need to build a host-nodejs package first, as there
are some tools needed to be executable on the host system
during the intermediate cross-build process.

This package drops support for NPM, as we will be managing
Node.js packages with Yarn, which will be implemented in the
following commits.

The patches that are shipped along with this version of the
package have been taken from the SRPM of the Node.js-package
in Fedora. See:

https://src.fedoraproject.org/rpms/nodejs/c/f0e2a1ed4615?branch=master

Signed-off-by: Björn Esser <b.esser@pengutronix.de>
[uol: fixed whitespace issues in first patch, added patch to get rid of the
non-existing rpath, dropped CREDITS lines, switched from menuconfig entry for
NODEJS back to config entry because of the removal of all sub-menu entries,
removed shell usage & inserted line break in build command]
Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>
Message-Id: <20200318155406.27584-1-u.oelmann@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>