host-yarn: new package
Fast, reliable, and secure dependency management for Node.js.

We use Yarn to generate an offline cache and a lock-file to be able to install
all needed Node.js-packages reproducibly. This works even in environments
without internet-connectivity, as long as the cache and the lock-file are in
reach.

Signed-off-by: Björn Esser <b.esser@pengutronix.de>
[uol: adjusted commit message, dropped CREDITS line, added Yarn's license,
wrapped install into call of world/execute and made it more verbose]
Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>
Message-Id: <20200318155406.27584-2-u.oelmann@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>