host-qemu: handle symlinks with 9p and security_model=mapped-file
Without this and security_model=mapped-file any symlink must be a regular
file and the symlink target should be inside the virtfs metadata.
With this patch regular symlinks are also accepted and created.

This makes it possible to use the same directory with
security_model=mapped-file, security_model=none and as a regular nfsroot.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>