improve rebasing for git-ptx-patches and multiple tags
git-ptx-patches needs the tags to match git commit to sections in the
series file. When rebasing across tags, those tags are lost and
git-ptx-patches does not work correctly any more.

With this change, the commit list for an interactive rebase is filtered
before the editor is opened. 'exec' lines are inserted in the appropriate
places to recreate the tags.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>