ptxd_make_pkghash: improve hash creation
The are some subtle issues with the code that are fixed here.

If multiple packages have the same patches, e.g. host and target, then the
patches where added twice. Deduplicate the directories to to avoid
searching twice.
Search with '-L' to follow any symlinks. We want the real content here.
Use the correct way to iterate over the list. Otherwise the order depends
on the awk version.
Don't search the current directory if no patches are found.
Don't sort the config files. The order is stable and sorting may changes
depending on the paths for PTXdist and BSP.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>