gst-validate1: fix gst-validate-launcher
Select Python SSL support, otherwise hashlib is not available and this is
needed to run gst-validate-launcher.

gst-validate-launcher looks at *.py files to find the 'test manager'.
This does not work if only *.pyc files are installed. So install *.py
instead for now. Without this, running gst-validate-launcher fails with:

Could not load testsuite: <module '...' from '...'> maybe because of
missing TestManager

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>