syslogng: update to 3.25.1
Update and move out of staging. The new version adds compatiblity with
OpenSSL 1.1.x and switches to Python 3.

The AMQP and MongoDB destinations are removed, they were never buildable
without additional packages not available in plain PTXdist. Various
other modules are disabled explicitly to avoid the build
nondeterministically picking up undeclared dependencies.

A simple replacement for the old systemd unit is added, as the new version
only provides an instanced unit which cannot be linked as syslog.service
(which is necessary to use syslog-ng as the default system logger by
enabling ForwardToSyslog in journald.conf).

Signed-off-by: Matthias Schiffer <matthias.schiffer@ew.tq-group.com>
Message-Id: <20200408123232.12718-1-matthias.schiffer@ew.tq-group.com>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>