wrapper: icecc: don't store the gcc switches in the debug section
By default a lot of gcc switches are stored in DW_AT_producer. This
includes '-fdirectives-only' and other options that are added by icecc. As
a result, the DW_AT_producer is different for locally and remotely compiled
objects. This produces a different build ID.

Add '-gno-record-gcc-switches' to the command-line when icecc is enabled.
In this case, only the compiler version is stored in DW_AT_producer.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>