u-boot-tools: version bump 2019.01 -> 2020.04
The new U-Boot version has a tools-only_defconfig for building the
tools. Switch from the sandbox_defconfig to the tools-only defconfig.

The U-Boot tools do not actually build the sandbox U-Boot and therefore
SDL support (for display/keyboard) is definitely not needed. Disable SDL
to prevent the "sdl2-config: Command not found" messages when building
the tools.

Use the documented envtools top-level command to build the target tools
instead of calling the Makefile in the tools/env directory.

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>
Message-Id: <20200414080249.8105-1-m.tretter@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>