ipkg-push: build index with sha256 checksums if supported
Instead of always building the index with the default md5sum, the index
will now be built with sha256, iff the target opkg package is
configured to support sha256.

Also, the ipkg support in ipkg-push has been removed, and it now always
uses the opkg tools instead. The name is kept, since everything else is
still using the 'ipkg' naming scheme though.
Message-Id: <20200417095414.9278-1-sfalsig@veritystudios.com>

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>