host-mfgtools: fix random extract failures
Without this, extracting may fail with:
mv: cannot move '.../mfgtools-1.3.167.UT1H1D/uuu_source-1.3.167/bzip2' to '.../mfgtools-1.3.167/bzip2':
Directory not empty

The problem is, that the tarball contains two toplevel directories with
overlapping subdirectories. During extraction, the two are moved ontop of
each other. This can fail depending on the order.

So just keep them next to each other to avoid this. Only one of them is
needed anyways. However, the .tarball-version file is in the wrong
directory, so copy it to the other one.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>