php5: version bump to 5.6.40
This is currently the last 5.x release of php, so this
version bump just closes known CVEs[1], and allow a little
extra time to migrate existing application to php 7.x.

Disabled fileinfo feature[2] as configure now contains a compile
check for existence of strcasestr. This fails as the cross
compiled binary cannot run on host arch.
This has an open upstream php bug[3].

New configuration parameters are default disabled.

[1] https://www.php.net/ChangeLog-5.php
[2] https://www.php.net/manual/en/book.fileinfo.php
[3] https://bugs.php.net/bug.php?id=69132

Signed-off-by: Bruno Thomsen <bruno.thomsen@gmail.com>
Message-Id: <20200429140452.23348-1-bruno.thomsen@gmail.com>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>