libseccomp: version bump 2.4.1 -> 2.4.3
Note: The new version is important for glibc 2.31:

Starting with this version, the clock_gettime() implementation calls the
clock_gettime64() syscall internaly.
The definition of this syscall was imported into libseccomp with the commit
"Update the syscall table for Linux v5.4-rc4" in version 2.4.2.

The current systemd version tries to whitelist clock_gettime64 but this is
ignored if libseccomp does not know about it.

As a result, with  glibc 2.31 and libseccomp 2.4.1 clock_gettime64 is not
allowed for any systemd service that uses SystemCallFilter= with a
whitelist.
Specifically systemd-journald, systemd-udevd and others don't run properly
and the system fails to boot.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>