rules: fix ptx/force-sh, ptx/sh handling with parallel building
When ptx/sh or ptx/force-sh is used in variables that are evaluated on
demand, then the modified $(SHELL) is actually visible to other commands.
Those will fail when they try to call bash functions.

So redefine ptx/force-sh and ptx/sh at the end of the toplevel makefile.
All ':=' variables that use it are already evaluated at this point.

This way $(SHELL) is only redefined during the single-threaded initial
evaluation and there are no visible side effects.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>