dtc: simplify dependency handling
Some of the dependencies where broken because of the variable evaluation
order. This was not noticed, because during dtb generation the same
dependencies where generated as well.
There is only one dependency not generated: When a dts is added in the
platformconfigdir that overwrites a dts in the kernel source tree.

This additional dependency is added here. So all other dependencies in the
makefile can be removed.
Also remove the dependency on kernel.extract.post. All dtbs depend on
dtc.install and dtc selects the kernel.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>