ptxd_lib_check_dir_permissions: fix permission check
The check fails if a package creates a directory with non-standard
permissions and then a file inside this directory.

To fix this, don't fail if the new (different) permissions are implicit but
the packages is the same.

Note: This also works when two packages install files into the same
directory with non-standard permissions:
For each package, the directory comes first. When the permission match,
then the package for the path is overwritten. When the parent directories
for the file are checked, then the package matches an no error occurs.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>