package templates: add code-signing-provider template
A ptxdist code signing provider is a package which selects the required
host tools needed for the code signing helpers to work. A shell script
is needed to define roles, set PKCS#11 URIs and import keys if SoftHSM
is used. In order to simplify its creation provide a template along with
an example script.

Signed-off-by: Bastian Krause <bst@pengutronix.de>
Message-Id: <20200617143125.23999-3-bst@pengutronix.de>
Tested-by: Roland Hieber <rhi@pengutronix.de>
Reviewed-by: Roland Hieber <rhi@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>