ptxd_template_write_src: print generated files with correct prefix
The second tar can only print the file names in the archive, and since
the first tar does not store the original path and since the files come
from the template path, we could not use that path anyway. Do some
additional handling to print the full paths of the generated files after
unpacking them to the destination.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>
Message-Id: <20200618113902.19225-1-rhi@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>