template: improve file template
This template is realy only there to install random files that don't need a
build stage. Usually config files etc.

So there is no need for a install stage. The long description is really
outdated anyways.

Don't comment out the targetinstall stage. It will be needed in most cases.
And use install_alternative as example. Package based on the 'file'
template don't have a <PKG>_DIR.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>