image-hdimg: add support for GPT
This adds an option to choose between GPT or MBR partition tables.
This ensures, that the correct partition type is set for all default
partitions.

This fixes a problem with the current vfat config and GPT:
Without this change both partition-type and partition-type-uuid are set.
With the current genimage version, this results in a 'hybrid' partition
table (both MBR and GPT). This can confuse BIOS and boot ROMs.

So explicitly choose between GPT and MBR and set the correct partition type
everywhere to avoid this.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>