kernel: cleanup and refactor KERNEL_MAKE_OPT and KERNEL_CONF_OPT
This introduces KERNEL_MODULE_OPT as a explicit variable that should be
used by kernel module packages. Hopefully, this will be the last time that
kernel module packages need to be changes due to changes in the kernel
rule. A new check ensures, that an error occurs when other packages use
KERNEL_MAKE_OPT.
cryptodev and the kernel module template now use KERNEL_MODULE_OPT.

Remove KERNEL_MODULES_BUILD. Use ptx/ifdef instead. That's much simpler and
avoids the extra kconfig option.

KERNEL_MAKE_OPT can now contain all necessary options so world/compile can
be used to build the kernel itself.

The variables are structured in a way that makes it possible to add extra
options to KERNEL_MAKE_OPT and KERNEL_CONF_OPT that are not used by the
other. This will be used in later commits.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>