kernel: add option to minimize changes to the kernelconfig
When the BSP contains a patch stack for the kernel, then it is often useful
to modify EXTRAVERSION with one of the patches. This makes it easier to
detect which version of the patch stack was used to build the currently
running kernel. However, this means that the comment at the beginning of
the kernel config changes with each new revision of the patch stack.

When this new option is enabled, then the version in the kernel config is
overwritten with the version from PTXCONF_KERNEL_VERSION. The version built
into the kernel remains unchanged.

This way the kernel config remains unchanged when the patches are updated.
This is especially useful with layers: If multiple layers change the kernel
config then this options avoids changes to the kernel config in all those
layers after changing the patches.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>