rootfs: don't create empty /media subdirs
If one of PTXCONF_ROOTFS_MEDIA* is "", then all later ones are not defined
at all. As a result '/media'. Is created multiple times with mode '0777'.
If PTXCONF_ROOTFS_MEDIA1 is not empty, then this implicitly creates
'/media' with mode '0755'. The result are conflicting permissions.

Don't create directories for empty PTXCONF_ROOTFS_MEDIA* to fix this.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>