networkmanager: version bump 1.22.10 -> 1.26.0 (latest stable)
There is a new option to install and use firewalld zone for shared mode;
we set it to false by default, until someone uses firewalld with
networkmanager.

Since 1.25.2-dev-90-g10020a946 in NetworkManager, there is a new tool
clients/cli/generate-docs-nm-settings-nmcli that generates the
nm-settings-docs-nmcli.xml based on nmcli meta data. Unfortunately, this
isn't built in meson's 'native' mode; as this needs a longer dependency
chain, disable generate-docs-nm-settings-nmcli when docs are disabled.

Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>
Message-Id: <20200725191307.2424653-3-r.schwebel@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>