kernel: make sure the deprecated error is always visible
For some reason when running 'ptxdist -j -q go', make aborts the build but
does not print the error message. If the dummy '$(shell :)' is there as
well, then the error is visible. It is not really clear, why this happens.
Just add it to avoid failures without error message.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>