zstd: remove libzstd targetinstall option
Drop useless ZSTD_LIBZSTD selection. The main purpose of the zstd
package is to install the libzstd. Explicit selecting the zstdlib is
not intuitive and causes errors if someone assumes that this lib gets
installed just by selection the zstd package.

Link the zstd helper against the shared libzstd while on it.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>
Message-Id: <20200812152703.7334-1-m.felsch@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>