ptxd_make_xpkg_pkg: make sure that directories have the correct permissions
ptxd_ensure_dir() is called for all parent directories when files are
installed. It uses 'install' to ensure that the parent directory exists. As
a result, the ownership and permissions of the directories are overwritten
with root:root / 0755.

Fix this by using 'mkdir -p' to create the parent directories. This works
correctly because ptxdist ensures now, that the umask is 0022. So the mode
will be 0755 for all new directories. Existing directories remain
unchanged and the ownership and permissions are preserved.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>