ptxd_lib_check_dir_permissions: handle corner cases better
Change the whole permission checking to validate the final permission each
directory should have in the opkg package.
This means that:
 - any explicitly created directory overwrites the ownership and
   permissions of any implicit creation
 - if a directory is created explicitly multiple times, then the last
   ownership and permissions are used.

To handle this, collect all the data for one package first and merge
multiple occurrences according to the rules specified above.

Then check and merge with the global database.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>