pkg-config-wrapper: fix a race condition
Currently, pkg-config-wrapper assumes, that if a pkg-config call fails
once, then it will fail again. So if pkg-config fails during the whitelist
check, it just continues and lets the next call to pkg-config produce the
error.

However, this is racy when packages are built in parallel: It can happen,
that the pkg-config whitelist call fails, because the package is not yet
available, and the following 'real' pkg-config call succeeds because the
.pc was installed in between the two calls.
As a result the whitelist check is bypassed. This usually leads to build
errors because followup pkg-config calls to query cflags or libs will be
blocked.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>