nvmetcli: new package
nvmetcli is a program used for viewing, editing, saving, and starting a
Linux kernel NVMe Target, used for an NVMe-over-Fabrics network
configuration. It allows an administrator to export a storage resource
(such as NVMe devices, files, and volumes) to a local block device and
expose them to remote systems based on the NVMe-over-Fabrics
specification from http://www.nvmexpress.org.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>
Message-Id: <20200921083057.4074995-2-l.stach@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>