git-ptx-patches: support flag -n to get unnumbered patches
When passing -n to git-ptx-patches the filenames in the patch stack
don't get a number as prefix.

The obvious downside is that the order of the patches isn't obvious
without the series file and you cannot create a patch stack if two
patches share have the same shortlog (which ideally shouldn't happen
...)

The obvious advantage is that when a patch is added or removed in the
middle of the series there are no renames for the later patches
necessary and the changes to series are smaller accordingly.

So if the patchstack of openssl used unnumbered patches the commit
c45d66cdae4b ("openssl: remove engines-path patch") would only have to
touch two lines in series (i.e. the removal of the patch and the changed
checksum) and contain no renames.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Message-Id: <20200922084730.11671-1-u.kleine-koenig@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>