icecc: handle older icecc versions
Older icecc versions can only add gcc or clang to the envronment.
So add the gcc arguments first. This way it will always be added but the
clang arguments may be ignored.
Check the resulting environment to see if clang is present.
Skip icecc in the wrapper if clang is not available.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>