ptxd_kconfig: handle BSP local toplevel Kconfig file correctly with layers
The toplevel Kconfig file for the ptxconfig and platformconfig is found
once at the beginning of ptxd_kconfig(). This can break oldconfig:
When iterating over layers, the Kconfig file found for the top layer is
always used. This is incorrect. Different layers may have a different
Kconfig file.

So update PTXDIST_PATH_LAYERS and look for the correct Kconfig file in
ptxd_kconfig_update() for each layer separately.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>