mtd-utils: spring cleanup
Drop unused config item MTD_UTILS_JITTERTEST which is unused since more
than 15 years.

Reactivate MTD_UTILS_JFFS2READER to install jffs2reader. This tool was
claimed to be removed upstream in commit ba0a8f624d02 (" * mtd-utils:
removed tools which are not in upstream any more"), which I failed to
find in the upstream git history.

mtdpart is a tool that was introduced in mtd-utils 1.5.2, add an
installation option for it.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Message-Id: <20210119144626.7368-1-u.kleine-koenig@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>