kconfig: update to v5.11-rc3
Rework the entire buildsystem. The Makefiles upstream has changed
completely so there is no need stick to anything that resembles the old
kernel Makefile. So clean it up and use autoconf to fill in the variables
from configure.

Use only pkg-config to find ncurses. This should work on all Linux
distributions. There was a report about problems with ncurses instead of
ncursesw, so use the later if it's available.

Drop anything qconf and gconf releated. We never built that anyways.

Always generate the parser code. Linux has been doing this for some time
now and it works just fine there.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>