ptxd_lib_dgen: provide lists for all existing and used but unknown kconfig symbols
There should be no unknown symbols. Kconfig accepts 'depends on UNKNOWN'
without warning or error. This is not desired in PTXdist. Such symbols with
show up in the unknown symbols list. The cannot be distinguished from
'range' limit numbers, so those show up as well in the raw data. So skip
all numbers.
The list of all PTXCONF_ symbols is also available. This can be used to
determine if PTXCONF_ variables used by makefiles actually exist.

This will be used by the new 'lint' command.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>